+++ CHANGELOG.md
@@ -30,6 +30,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   glob patterns. This can used from the `cached-path` command with `allennlp cached-path --remove some-files-*`.
 - Added logging for the main process when running in distributed mode.
 - Added a `TrainerCallback` object to support state sharing between batch and epoch-level training callbacks.
+- Added classes: `nn/samplers/samplers.py` with `MultinomialSampler`, `TopKSampler`, and `TopPSampler` for 
+  sampling indices from log probabilities
+- Made `BeamSearch` registrable.
+- Added `top_k_sampling` and `type_p_sampling` `BeamSearch` implementations.
 
 ### Changed